refacto 3
