+++ src-tauri/src/fast_insert.rs
@@ -1,6 +1,6 @@
 use std::fs::File;
 use csv::{Reader, StringRecord};
-use crate::{DatabaseConnection};
+use crate::fileflow::database_connection::DatabaseConnection;
 
 use crate::fileflow::fileflow::{get_drop_statement, get_create_statement, get_insert_into_statement};
