+++ src-tauri/src/fileflow/database_connection.rs
@@ -1,4 +1,4 @@
-use sqlx::{Error, MySqlPool, PgPool, Pool, Sqlite, SqlitePool};
+use sqlx::{Error, MySqlPool, PgPool, SqlitePool};
 
 use crate::fileflow::db_config::DbConfig;
 
@@ -10,31 +10,28 @@ pub enum DatabaseConnection {
 
 impl DatabaseConnection {
     pub async fn connect(config: &DbConfig) -> Result<DatabaseConnection, Error> {
+        let connection_str = match config.db_driver.as_str() {
+            "postgres" => format!(
+                "postgres://{}{}@{}:{}/{}",
+                config.username,
+                if config.password.is_empty() { "".to_string() } else { format!(":{}", config.password) },
+                config.db_host, config.port, config.db_name
+            ),
+            "mysql" | "mariadb" => format!(
+                "mysql://{}{}@{}:{}/{}",
+                config.username,
+                if config.password.is_empty() { "".to_string() } else { format!(":{}", config.password) },
+                config.db_host, config.port, config.db_name
+            ),
+            "sqlite" => config.sqlite_file_path.clone(),
+            _ => return Err(Error::Protocol(format!("Unsupported database driver: {}", config.db_driver).into())),
+        };
+
         match config.db_driver.as_str() {
-            "postgres" => {
-                let pool = PgPool::connect(&format!(
-                    "postgres://{}:{}@{}:{}/{}",
-                    config.username, config.password, config.db_host, config.port, config.db_name
-                ))
-                    .await?;
-                Ok(DatabaseConnection::Postgres(pool))
-            }
-            "mysql" | "mariadb" => {
-                let pool = MySqlPool::connect(&format!(
-                    "mysql://{}:{}@{}:{}/{}",
-                    config.username, config.password, config.db_host, config.port, config.db_name
-                ))
-                    .await?;
-                Ok(DatabaseConnection::MySQL(pool))
-            }
-            "sqlite" => {
-                let pool: Pool<Sqlite> = SqlitePool::connect(&config.sqlite_file_path).await?;
-                Ok(DatabaseConnection::SQLite(pool))
-            }
-            _ => Err(Error::Protocol(format!(
-                "Unsupported database driver: {}",
-                config.db_driver
-            ).into())),
+            "postgres" => PgPool::connect(&connection_str).await.map(DatabaseConnection::Postgres),
+            "mysql" | "mariadb" => MySqlPool::connect(&connection_str).await.map(DatabaseConnection::MySQL),
+            "sqlite" => SqlitePool::connect(&connection_str).await.map(DatabaseConnection::SQLite),
+            _ => unreachable!(),
         }
     }
