+++ src-tauri/src/fileflow/mod.rs
@@ -3,3 +3,4 @@ pub(crate) mod db_config;
 pub(crate) mod insert_config;
 pub(crate) mod save_config;
 pub(crate) mod database_connection;
+pub mod action;
